# src/main/resources/static/openapi/SENTINEL_PoC_API.yaml
...
openapi: 3.0.3
info:
  title: SENTINEL PoC API
  version: 0.1.0
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: 로그인 (토큰 발급)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthLogin'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: 로그아웃 (토큰 무효화)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
  # 기업 추가 페이지: 기업명/기업코드 검색
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiResponseAuthLogin:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/AuthLoginResponse'
        error:
          nullable: true
    ApiResponseMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
        error:
          nullable: true
...
    UserRefDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          nullable: true
          enum:
            - USER
            - ADMIN
    AuthLoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: seconds
        user:
          $ref: '#/components/schemas/UserRefDto'
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
...